+++ src/App.tsx
@@ -1,3 +1,20 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { useEffect } from "react";
 import "./css/App.css";
 import "@fontsource/ibm-plex-sans";
@@ -39,11 +56,18 @@ function App() {
 
     useEffect(() => {
         if (user && !session) {
-            api.meSession()
+            api.meSessionCurrent()
                 .then(setSession)
                 .catch((data: ErrorDto) => {
                     if (data.statusCode === 401) {
                         localStorage.removeItem("token");
+                        if ("serviceWorker" in navigator) {
+                            navigator.serviceWorker.ready.then(async (registration) => {
+                                const subscription =
+                                    await registration.pushManager.getSubscription();
+                                subscription?.unsubscribe();
+                            });
+                        }
                         return window.location.reload();
                     } else {
                         setNotification({