+++ src/components/conversation/comment/player.tsx
@@ -1,3 +1,20 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { Box, IconButton, Tooltip } from "@mui/material";
 import React, { useRef, useState } from "react";
 import YoutubePlayer from "react-player/youtube";
@@ -28,8 +45,8 @@ export default function Player(props: { url: string; style?: React.CSSProperties
     const { url, style } = props;
 
     const mode =
-        (regex.facebook.videos.some((regexp) => url.match(regexp)) && "facebook") ||
-        (regex.youtube.some((regexp) => url.match(regexp)) && "youtube") ||
+        (regex.facebook.videos.some((regexp) => regexp.test(url)) && "facebook") ||
+        (regex.youtube.some((regexp) => regexp.test(url)) && "youtube") ||
         "streamable";
 
     const buttons = [