+++ src/components/conversation/comment/voteButtons.tsx
@@ -1,3 +1,20 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import React, { useState } from "react";
 import { ArrowDropDown, ArrowDropUp } from "@mui/icons-material";
 import { Button, ButtonGroup, Typography } from "@mui/material";
@@ -42,7 +59,7 @@ export default function VoteButtons(props: { comment: Comment }) {
         <ButtonGroup variant="text" className="!rounded-[4px] !bg-[#333]">
             <Button
                 className="!p-0 !m-0 !block !py-[2px] !min-w-0 !pl-[5.5px] !pr-[6px] !rounded-r-0"
-                disabled={!user || !!vote}
+                disabled={!user || Boolean(vote)}
                 onClick={() => {
                     sendVote("U");
                 }}
@@ -59,7 +76,7 @@ export default function VoteButtons(props: { comment: Comment }) {
             </Button>
             <Button
                 className="!p-0 !m-0 !block !py-[2px] !min-w-0 !pr-[10px] !rounded-l-0"
-                disabled={!user || !!vote}
+                disabled={!user || Boolean(vote)}
                 onClick={() => {
                     sendVote("D");
                 }}