+++ src/lib/icons/dicord.tsx
@@ -1,4 +1,20 @@
 /*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+/*
 Jsr-mandarine, CC BY-SA 4.0 <https://creativecommons.org/licenses/by-sa/4.0>, via Wikimedia Commons
 */
 import React from "react";
@@ -43,7 +59,7 @@ export default function DiscordIcon(props: {
                         (white && "#ffffff") || (black && "#000000") || color || "#5865F2"
                     }
                     fillRule="nonzero"
-                ></path>
+                />
             </g>
         </svg>
     );