+++ src/lib/popup.tsx
@@ -1,3 +1,20 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import React from "react";
 import { Close } from "@mui/icons-material";
 import {
@@ -80,7 +97,7 @@ export function PopUp(props: {
                 >
                     {children}
                 </Box>
-                {!!buttons?.length && (
+                {Boolean(buttons?.length) && (
                     <React.Fragment>
                         <Divider />
                         <Box className="flex w-full">